+++ Runtime/Graphics/Graphics.cs
@@ -1,0 +1,10 @@
+namespace Voxell.Graphics
+{
+  public static class Graphics
+  {
+    public static int S_BLOCK_SZ = 32;
+    public static int M_BLOCK_SZ = 64;
+    public static int L_BLOCK_SZ = 128;
+    public static int XL_BLOCK_SZ = 512;
+  }
+}