+++ Tests/Editor/JobxTests.cs
@@ -1,23 +1,27 @@
 using Unity.Mathematics;
 using Unity.Collections;
 using NUnit.Framework;
-using Voxell.Mathx;
 using Random = UnityEngine.Random;
 
 namespace Voxell.Jobx
 {
+  using Mathx;
+
   public class JobxTests
   {
-    private const int ARRAY_COUNT = 1000;
+    private const int ARRAY_COUNT = 2000;
+
     [Test]
-    public void SumScan()
+    public void SumScanJobTest()
     {
       int[] array = new int[ARRAY_COUNT];
       for (int i=0; i < ARRAY_COUNT; i++) array[i] = GenerateRandomInt();
 
       NativeArray<int> na_array = new NativeArray<int>(array, Allocator.TempJob);
-      Jobx.InclusiveSumScan(na_array);
+      SumScanJob sumScanJob = new SumScanJob(ref na_array);
+      sumScanJob.InclusiveSumScan();
 
+      // using serial inclusive sum scan method to make sure that the parallel method works
       int sum = 0;
       for (int i=0; i < ARRAY_COUNT; i++)
       {
@@ -26,17 +30,20 @@ namespace Voxell.Jobx
       }
 
       na_array.Dispose();
+      sumScanJob.Dispose();
     }
 
     [Test]
-    public void Float3MinScan()
+    public void Float3MinScanJobTest()
     {
       float3[] array = new float3[ARRAY_COUNT];
       for (int i=0; i < ARRAY_COUNT; i++) array[i] = GenerateRandomFloat3();
 
       NativeArray<float3> na_array = new NativeArray<float3>(array, Allocator.TempJob);
-      Jobx.InclusiveFloat3MinScan(na_array);
+      Float3MinScanJob float3MinScanJob = new Float3MinScanJob(ref na_array);
+      float3MinScanJob.InclusiveMinScan();
 
+      // using serial inclusive min scan method to make sure that the parallel method works
       float3 float3Max = array[0];
       for (int i=0; i < ARRAY_COUNT; i++)
       {
@@ -45,17 +52,20 @@ namespace Voxell.Jobx
       }
 
       na_array.Dispose();
+      float3MinScanJob.Dispose();
     }
 
     [Test]
-    public void Float3MaxScan()
+    public void Float3MaxScanJobTest()
     {
       float3[] array = new float3[ARRAY_COUNT];
       for (int i=0; i < ARRAY_COUNT; i++) array[i] = GenerateRandomFloat3();
 
       NativeArray<float3> na_array = new NativeArray<float3>(array, Allocator.TempJob);
-      Jobx.InclusiveFloat3MaxScan(na_array);
+      Float3MaxScanJob float3MaxScanJob = new Float3MaxScanJob(ref na_array);
+      float3MaxScanJob.InclusiveMaxScan();
 
+      // using serial inclusive max scan method to make sure that the parallel method works
       float3 float3Max = array[0];
       for (int i=0; i < ARRAY_COUNT; i++)
       {
@@ -64,10 +74,11 @@ namespace Voxell.Jobx
       }
 
       na_array.Dispose();
+      float3MaxScanJob.Dispose();
     }
 
     [Test]
-    public void RadixSort()
+    public void RadixSortJobTest()
     {
       uint[] array = new uint[ARRAY_COUNT];
       int[] indices = MathUtil.GenerateSeqArray(ARRAY_COUNT);
@@ -75,7 +86,8 @@ namespace Voxell.Jobx
 
       NativeArray<uint> na_values = new NativeArray<uint>(array, Allocator.TempJob);
       NativeArray<int> na_indices = new NativeArray<int>(indices, Allocator.TempJob);
-      Jobx.RadixSort(na_values, na_indices);
+      RadixSortJob radixSortJob = new RadixSortJob(ref na_values, ref na_indices);
+      radixSortJob.Sort();
 
       // check if sorting works
       for (int i=0; i < ARRAY_COUNT-1; i++)
@@ -86,10 +98,11 @@ namespace Voxell.Jobx
 
       na_values.Dispose();
       na_indices.Dispose();
+      radixSortJob.Dispose();
     }
 
-    private float3 GenerateRandomFloat3() => Random.insideUnitSphere * Random.Range(0.0f, 100.0f);
-    private int GenerateRandomInt() => Random.Range(0, 100);
-    private uint GenerateRandomUInt() => (uint)Random.Range(0, 100);
+    private float3 GenerateRandomFloat3() => Random.insideUnitSphere * Random.Range(0.0f, (float)ARRAY_COUNT);
+    private int GenerateRandomInt() => Random.Range(0, ARRAY_COUNT);
+    private uint GenerateRandomUInt() => (uint)Random.Range(0, ARRAY_COUNT);
   }
 }